+++ src/components/ConversationEnded.tsx
@@ -6,7 +6,6 @@ import { Flex } from "@twilio-paste/core/flex";
 import { Text } from "@twilio-paste/core/text";
 import { Button } from "@twilio-paste/core/button";
 import { useDispatch, useSelector } from "react-redux";
-import { Media, Message, User } from "@twilio/conversations";
 import { useState } from "react";
 import slugify from "slugify";
 
@@ -14,125 +13,13 @@ import { sessionDataHandler, contactBackend } from "../sessionDataHandler";
 import { changeEngagementPhase, updatePreEngagementData } from "../store/actions/genericActions";
 import { EngagementPhase, AppState } from "../store/definitions";
 import { containerStyles, textStyles, titleStyles } from "./styles/ConversationEnded.styles";
-import { generateDuration } from "../utils/generateDuration";
-
-export interface Transcript {
-    author?: string;
-    body: string;
-    timeStamp: Date;
-    attachedMedia?: Media[] | null;
-}
-
-export const getTranscriptData = (messages: Message[] | undefined, users: User[] | undefined): Transcript[] => {
-    const transcriptData = [];
-    if (messages && users) {
-        for (const message of messages) {
-            const currentUser = users.find((user) => user.identity === message.author);
-            transcriptData.push({
-                author: message.author === "Concierge" ? message.author : currentUser?.friendlyName,
-                body: message.body,
-                timeStamp: message.dateCreated,
-                attachedMedia: message.attachedMedia
-            });
-        }
-    }
-    return transcriptData;
-};
-
-export const getAgentNames = (customerName: string | undefined, transcriptData: Transcript[]) => {
-    const names = transcriptData.map((message) => message.author);
-    let agentNames = Array.from(
-        new Set(names.filter((name) => name?.trim() !== customerName && name?.trim() !== "Concierge"))
-    );
-    agentNames = agentNames.map((name) => name?.trim());
-    return agentNames;
-};
-
-const getUniqueFilenames = (transcriptData: Transcript[]) => {
-    const mediaMessages = transcriptData.filter((message) => message.attachedMedia);
-    const filenames = mediaMessages.map((message) => {
-        if (message.attachedMedia) {
-            return message.attachedMedia[0].filename;
-        }
-        return "";
-    });
-    interface seenFilenamesInfo {
-        [key: string]: number;
-    }
-    const seenFilenames: seenFilenamesInfo = {};
-    const uniqueFilenames = [];
-    for (const filename of filenames) {
-        if (Object.keys(seenFilenames).includes(filename)) {
-            const filenameParts = filename.split(".");
-            filenameParts[0] = `${filenameParts[0]}-${seenFilenames[filename]}`;
-            uniqueFilenames.push(filenameParts.join("."));
-            seenFilenames[filename] += 1;
-        } else {
-            seenFilenames[filename] = 1;
-            uniqueFilenames.push(filename);
-        }
-    }
-    return uniqueFilenames;
-};
-
-export const generateDownloadTranscript = (
-    customerName: string | undefined,
-    agentNames: (string | undefined)[],
-    transcriptData: Transcript[]
-) => {
-    const doubleDigit = (number: number) => `${number < 10 ? 0 : ""}${number}`;
-    const conversationStartDate = transcriptData[0].timeStamp.toLocaleString("default", { dateStyle: "long" });
-    const duration = generateDuration(transcriptData);
-    const uniqueFilenames = getUniqueFilenames(transcriptData);
-    let mediaMessageIndex = 0;
-    let conversationTitle = `Conversation with ${customerName}`;
-    if (agentNames.length > 0) {
-        agentNames.forEach((name) => (conversationTitle = conversationTitle.concat(` and ${name}`)));
-    }
-    let transcript = `${conversationTitle}\n\nDate: ${conversationStartDate}\nDuration: ${duration}\n\n`;
-    for (const message of transcriptData) {
-        const bulletPoint = message.author === customerName ? "*" : "+";
-        let messageText = `${bulletPoint} ${doubleDigit(message.timeStamp.getHours())}:${doubleDigit(
-            message.timeStamp.getMinutes()
-        )}  ${message.author}: ${message.body}`;
-        if (message.attachedMedia) {
-            messageText = messageText.concat(` (** Attached file ${uniqueFilenames[mediaMessageIndex]} **)`);
-            mediaMessageIndex += 1;
-        }
-        transcript = transcript.concat(`${messageText}\n\n`);
-    }
-    return transcript;
-};
-
-export const generateEmailTranscript = (
-    customerName: string | undefined,
-    agentNames: (string | undefined)[],
-    transcriptData: Transcript[]
-) => {
-    const doubleDigit = (number: number) => `${number < 10 ? 0 : ""}${number}`;
-    const conversationStartDate = transcriptData[0].timeStamp.toLocaleString("default", { dateStyle: "long" });
-    const duration = generateDuration(transcriptData);
-    const uniqueFilenames = getUniqueFilenames(transcriptData);
-    let mediaMessageIndex = 0;
-
-    let conversationTitle = `Chat with <strong>${customerName}</strong>`;
-    if (agentNames.length > 0) {
-        agentNames.forEach((name) => (conversationTitle = conversationTitle.concat(` and <strong>${name}</strong>`)));
-    }
-
-    let transcript = `${conversationTitle}<br><br><strong>Date:</strong> ${conversationStartDate}<br><strong>Duration:</strong> ${duration}<br><br>`;
-    for (const message of transcriptData) {
-        let messageText = `${doubleDigit(message.timeStamp.getHours())}:${doubleDigit(
-            message.timeStamp.getMinutes()
-        )} <i>${message.author}</i>: ${message.body}`;
-        if (message.attachedMedia) {
-            messageText = messageText.concat(` (** Attached file <i>${uniqueFilenames[mediaMessageIndex]}</i> **)`);
-            mediaMessageIndex += 1;
-        }
-        transcript = transcript.concat(`${messageText}<br><br>`);
-    }
-    return transcript;
-};
+import {
+    getTranscriptData,
+    getAgentNames,
+    generateDownloadTranscript,
+    getUniqueFilenames,
+    generateEmailTranscript
+} from "./GenerateTranscripts";
 
 export const ConversationEnded = () => {
     const dispatch = useDispatch();
